consolidate pca9585 driver and hal adapter in single package. fix switch off issue at 100%
