+++ pca9685/pca9685.go
@@ -1,9 +1,9 @@
-package drivers
+package pca9685
 
 import (
+	"log"
 	"math"
 	"time"
-	"log"
 
 	"github.com/reef-pi/rpi/i2c"
 )
@@ -23,7 +23,7 @@ type PCA9685 struct {
 	Freq int
 }
 
-func NewPCA9685(addr byte, bus i2c.Bus) *PCA9685 {
+func New(addr byte, bus i2c.Bus) *PCA9685 {
 	return &PCA9685{
 		addr: addr,
 		bus:  bus,
@@ -51,7 +51,7 @@ func (p *PCA9685) Sleep() error {
 	if err != nil {
 		return err
 	}
-	
+
 	sleepmode := (mode1Reg & 0x7F) | 0x10 // Mask restart bit and set sleep bit
 	return p.bus.WriteToReg(p.addr, mode1RegAddr, []byte{sleepmode})
 }
@@ -62,7 +62,7 @@ func (p *PCA9685) Wake() error {
 	if err != nil {
 		return err
 	}
-	
+
 	if (mode1Reg & 0x80) != 0 {
 		// We are in sleep mode after a previous run without shutdown. Restore.
 		// First, clear sleep bit
@@ -86,10 +86,10 @@ func (p *PCA9685) Wake() error {
 	if err := p.bus.WriteToReg(p.addr, preScaleRegAddr, []byte{preScaleValue}); err != nil {
 		return err
 	}
-	
+
 	// Set our operating modes:
 	mode1Reg = 0x20 // No AllCall, no subaddresses, no sleep, internal clock, enable auto increment
-	
+
 	return p.bus.WriteToReg(p.addr, mode1RegAddr, []byte{mode1Reg})
 }
 
@@ -99,40 +99,38 @@ func (p *PCA9685) SetPwm(channel int, onTime, offTime uint16) error {
 	// The PCA9685 has two special states, full on and full off, besides the normal PWM.
 	// Using them prevents the microspikes that can cause extra heat generation in mosfet
 	// output stages as well as switching noise.
-	// Generally, if onTime + 1 == offTime, we're dealing with full on. If onTime == offTome, 
+	// Generally, if onTime + 1 == offTime, we're dealing with full on. If onTime == offTime,
 	// it's full off.
 	// Since onTime is 0 and always be 0, and offTime will vary between 0 .. 4095
 	// , we can use that as an indicator.
 	// 100 * 40.95 will result in 4095. Sanity check it anyway.
-	if offTime >= 4095 {
+	if offTime > 4095 {
 		offTime = 4095
-	} 
-	
+	}
+
 	// If offTime == 0, we want to be full off. Set LEDx_OFF_H(4)
 	if offTime == 0 {
-		offTime = 4096
-		onTime = 0
+		onTime = 4096 // check with onTimne 0 and off time 4096
 	}
-	
+
 	// If offTime == 4095, we want to be full on. Set LEDx_ON_H(4)
 	if offTime == 4095 {
-		onTime = 4096
-		offTime = 0
-	} 
+		onTime = 0 // check with setting off to 0 and on to 4096
+	}
 
-	// Split the ints into 4 bytes	
+	// Split the ints into 4 bytes
 	timeReg := byte(pwm0OnLowReg + (4 * channel))
 	onTimeLow := byte(onTime & 0xFF)
 	onTimeHigh := byte(onTime >> 8)
 	offTimeLow := byte(offTime & 0xFF)
 	offTimeHigh := byte(offTime >> 8)
-	
+
 	log.Println("onLow ", onTimeLow, " onHigh ", onTimeHigh, " offLow ", offTimeLow, " offHigh ", offTimeHigh)
 	// Send one entire channel in one go
 	if err := p.bus.WriteToReg(p.addr, timeReg, []byte{onTimeLow, onTimeHigh}); err != nil {
 		return err
 	}
-	return p.bus.WriteToReg(p.addr, timeReg + 2, []byte{offTimeLow, offTimeHigh})
+	return p.bus.WriteToReg(p.addr, timeReg+2, []byte{offTimeLow, offTimeHigh})
 }
 
 func (p *PCA9685) Close() error {
